Reprint Server WordPress Plugin: Fixups after latest trunk changes landed
