+++ packages/reprint-client/src/import.php
@@ -1174,7 +1174,7 @@ class ImportClient
 
         $this->initialize_tuner($options);
 
-        // Initialize HMAC authentication if a shared secret was provided.
+        // Initialize HMAC authentication if a connection token was provided.
         // When set, every outgoing HTTP request will include X-Auth-Signature,
         // X-Auth-Nonce, and X-Auth-Timestamp headers so the export API can verify
         // the caller without a SECRET_KEY in the URL.
@@ -1642,7 +1642,7 @@ class ImportClient
      * @param array $options {
      *     Parsed files-push options and context.
      *
-     *     @type string $secret             HMAC shared secret.
+     *     @type string $secret             HMAC connection token.
      *     @type bool   $force_http         Whether the operator allowed a plain-HTTP target.
      *     @type string $progress           Progress output mode: auto, tty, or jsonl.
      *     @type array  $files_push_context Optional context already validated by the CLI entry point.
@@ -2113,7 +2113,7 @@ class ImportClient
      * @param array $options {
      *     Parsed files-push options.
      *
-     *     @type string $secret     HMAC shared secret.
+     *     @type string $secret     HMAC connection token.
      *     @type bool   $force_http Whether the operator allowed a plain-HTTP target.
      * }
      * @phpstan-param array<string,mixed> $options
@@ -11508,8 +11508,8 @@ class ImportClient
                     'message' =>
                         "No --secret was provided. The remote site requires " .
                         "authentication.\n\n" .
-                        "Pass --secret=YOUR_SECRET using the same secret " .
-                        "configured under Tools > Reprint Server on the remote site.",
+                        "Pass --secret=YOUR_SECRET using the same connection " .
+                        "token configured under Tools > Reprint Server on the remote site.",
                 ];
             }
 
@@ -11532,7 +11532,7 @@ class ImportClient
                 return [
                     'code' => 'AUTH_SECRET_MISMATCH',
                     'message' =>
-                        "Wrong shared secret. The --secret value does not match " .
+                        "Wrong connection token. The --secret value does not match " .
                         "the one configured under Tools > Reprint Server in wp-admin.",
                 ];
             }
@@ -12773,7 +12773,7 @@ if (
             'type' => 'value',
             'target' => 'secret',
             'placeholder' => 'TOKEN',
-            'help' => 'HMAC shared secret for export API authentication',
+            'help' => 'HMAC connection token for export API authentication',
             'help_section' => 'global',
             'commands' => ['pull', 'pull-files', 'pull-db', 'files-pull', 'files-push', 'files-index', 'db-pull', 'db-index', 'preflight', 'preflight-assert'],
         ],
